+++ src/ondusAppliance.ts
@@ -1,19 +1,26 @@
 import { PlatformAccessory, Service } from 'homebridge';
 
+import { PLUGIN_VERSION } from './settings';
 import { OndusPlatform } from './ondusPlatform';
 import { OndusThresholds } from './ondusThresholds';
 import { OndusNotification, NOTIFICATION_CATEGORY_CRITICAL } from './ondusNotification';
-
 
 /**
  * Platform Accessory
  * An instance of this class is created for each accessory your platform registers
  * Each accessory may expose multiple services of different service types.
  */
+
 export abstract class OndusAppliance {
   static ONDUS_TYPE = 0;
   static ONDUS_PROD = 'Grohe AG'
-  static ONDUS_NAME = 'Overload me';
+  static ONDUS_NAME = 'Abstract';
+  // Hacky workaround for lack of reflection support in typescript
+  private ONDUS_MAP = {
+    101 : 'Sense',
+    102 : 'Sense Plus',
+    103 : 'Sense Guard',
+  }
 
   logPrefix: string;
   applianceID: string;
@@ -50,7 +57,19 @@ export abstract class OndusAppliance {
     // Update configured threshold limits
     this.thresholds.update();
 
+    // set accessory information
+    const ondusType = accessory.context.device.type;
+    this.accessory.getService(this.ondusPlatform.Service.AccessoryInformation)!
+      .setCharacteristic(this.ondusPlatform.Characteristic.Manufacturer, OndusAppliance.ONDUS_PROD)
+      .setCharacteristic(this.ondusPlatform.Characteristic.Model, this.ONDUS_MAP[ondusType])
+      .setCharacteristic(this.ondusPlatform.Characteristic.Name, accessory.context.device.name)
+      .setCharacteristic(this.ondusPlatform.Characteristic.HardwareRevision, accessory.context.device.type)
+      .setCharacteristic(this.ondusPlatform.Characteristic.SerialNumber, this.unhexlify(accessory.context.device.serial_number))
+      .setCharacteristic(this.ondusPlatform.Characteristic.FirmwareRevision, accessory.context.device.version)
+      .setCharacteristic(this.ondusPlatform.Characteristic.SoftwareRevision, PLUGIN_VERSION)
+      .setCharacteristic(this.ondusPlatform.Characteristic.AppMatchingIdentifier, 'id1451814256');
 
+    
     // Initialize common sensor services
 
     /**
@@ -63,8 +82,9 @@ export abstract class OndusAppliance {
     this.leakService
       .setCharacteristic(this.ondusPlatform.Characteristic.Name, accessory.context.device.name)
       .setCharacteristic(this.ondusPlatform.Characteristic.LeakDetected, this.ondusPlatform.Characteristic.LeakDetected.LEAK_NOT_DETECTED)
-      .setCharacteristic(this.ondusPlatform.Characteristic.StatusFault, this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
-
+      .setCharacteristic(this.ondusPlatform.Characteristic.Active, this.ondusPlatform.Characteristic.Active.ACTIVE)
+      .setCharacteristic(this.ondusPlatform.Characteristic.StatusFault, this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);      
+    
     // create handlers for required characteristics of Leak service
     this.leakService.getCharacteristic(this.ondusPlatform.Characteristic.LeakDetected)
       .on('get', this.handleLeakDetectedGet.bind(this));
@@ -80,8 +100,9 @@ export abstract class OndusAppliance {
     // set the Temperature service characteristics
     this.temperatureService
       .setCharacteristic(this.ondusPlatform.Characteristic.Name, accessory.context.device.name)
+      .setCharacteristic(this.ondusPlatform.Characteristic.Active, this.ondusPlatform.Characteristic.Active.ACTIVE)
       .setCharacteristic(this.ondusPlatform.Characteristic.StatusFault, this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
-    
+
     // create handlers for required characteristics of Temperature service
     this.temperatureService.getCharacteristic(this.ondusPlatform.Characteristic.CurrentTemperature)
       .on('get', this.handleCurrentTemperatureGet.bind(this));
@@ -99,6 +120,60 @@ export abstract class OndusAppliance {
   start(): void {
     return; 
   }
+
+
+  // ---- CHARACTERISTICS HANDLER FUNCTIONS BELOW ----
+
+  setLeakServiceLeakDetected(active: boolean) {
+    if (active) {
+      this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.LeakDetected, 
+        this.ondusPlatform.Characteristic.LeakDetected.LEAK_DETECTED);
+    } else {
+      this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.LeakDetected, 
+        this.ondusPlatform.Characteristic.LeakDetected.LEAK_NOT_DETECTED);
+    }
+  }
+
+  setLeakServiceActive(active: boolean) {
+    if (active) {
+      this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
+        this.ondusPlatform.Characteristic.Active.ACTIVE);
+    } else {
+      this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
+        this.ondusPlatform.Characteristic.Active.INACTIVE);      
+    }
+  }
+
+  setLeakServiceStatusFault(active: boolean) {
+    if (active) {
+      this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
+        this.ondusPlatform.Characteristic.StatusFault.GENERAL_FAULT);
+    } else {
+      this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusActive, 
+        this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);      
+    }
+  }
+
+  setTemperatureServiceActive(active: boolean) {
+    if (active) {
+      this.temperatureService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
+        this.ondusPlatform.Characteristic.Active.ACTIVE);
+    } else {
+      this.temperatureService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
+        this.ondusPlatform.Characteristic.Active.INACTIVE);      
+    }
+  }
+
+  setTemperatureServiceStatusFault(active: boolean) {
+    if (active) {
+      this.temperatureService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
+        this.ondusPlatform.Characteristic.StatusFault.GENERAL_FAULT);
+    } else {
+      this.temperatureService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
+        this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);      
+    }
+  }
+
 
   // ---- HTTP HANDLER FUNCTIONS BELOW ----
 
@@ -150,17 +225,10 @@ export abstract class OndusAppliance {
         });
 
         // Update the leakService LeakDetected characteristics
-        if (this.leakDetected) {
-          this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.LeakDetected, 
-            this.ondusPlatform.Characteristic.LeakDetected.LEAK_DETECTED);
-        } else {
-          this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.LeakDetected, 
-            this.ondusPlatform.Characteristic.LeakDetected.LEAK_NOT_DETECTED);
-        }
-        // Reset StatusFault characteristics for battery service
-        this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
-          this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
-    
+        this.setLeakServiceLeakDetected(this.leakDetected);
+        
+        // Enable Active characteristics for leak service
+        this.setLeakServiceActive(true);
         callback(null, this.leakDetected);
 
       })
@@ -168,10 +236,8 @@ export abstract class OndusAppliance {
         this.ondusPlatform.log.debug(err);
         this.ondusPlatform.log.error(`[${this.logPrefix}] Unable to process notifications: ${err}`);
         
-        // Set StatusFault characteristics for leakage service
-        this.leakService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
-          this.ondusPlatform.Characteristic.StatusFault.GENERAL_FAULT);  
-  
+        // Disable Active characteristics for leakage service
+        this.setLeakServiceActive(false);  
         callback(err, this.leakDetected);
       });
   }
@@ -237,7 +303,7 @@ export abstract class OndusAppliance {
         this.thresholds.update();
       })
       .catch( err => {
-        this.ondusPlatform.log.error(`[${this.logPrefix}] Unable to update appliance info: ${err.text}`);
+        this.ondusPlatform.log.error(`[${this.logPrefix}] Unable to update appliance info: ${err}`);
       });
   }