+++ src/settings.ts
@@ -7,3 +7,10 @@ export const PLATFORM_NAME = 'Ondus';
  * This must match the name of your plugin as defined the package.json
  */
 export const PLUGIN_NAME = 'homebridge-grohe-sense';
+
+/**
+ * Version number from package.json. 
+ * 
+ * WARNING: Must be updated manually before release!
+ */
+export const PLUGIN_VERSION = '1.3.1';