+++ CHANGELOG.md
@@ -5,15 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Released]
 
-## [1.X.X] - 
+## [1.3.2] - 2020-08-28 
 
 ### Added
 
+- A new Ondus API notification (category,type) message that was previously missing
+
 ### Changed
 
-## [Released]
+- Fixed a bug where status fault characteristics were never cleared after a message in Ondus API was marked as read
 
 ## [1.3.1] - 2020-08-25